backend/requirements.txt
flask
flask-socketio
kafka-python
eventlet
pymongo
werkzeug
flask-cors
confluent-kafka